test(packages/eg-walker): dedupe shuffleWithSeed into shared shuffle.ts
Five property test files each carried an identical Fisher–Yates shuffle
keyed on a 32-bit seed, with the same Numerical Recipes LCG constants
inlined four times. Extracts a single shared `./shuffle.ts` so the
constants live in one place and shrunk seeds keep their semantics if
the LCG ever changes.

Also clarifies the `eventDagArb` docstring: non-root nodes get 1–2
parents (not 0–2), because `missing-parent-buffering` relies on the
DAG being connected to exercise the buffer flush path.
