+++ packages/eg-walker/src/test/property/delivery-order-invariance.property.test.ts
@@ -12,29 +12,8 @@ import { EgWalkerReplica } from "../../core/replica";
 import { cloneEvent } from "../test-helpers";
 import { traceParamsArb } from "./arbitraries";
 import { fcParams } from "./run-config";
+import { shuffleWithSeed } from "./shuffle";
 import { runTrace } from "./trace-runner";
-
-/**
- * Fisher–Yates shuffle keyed on a 32-bit seed via the same LCG used in
- * `test-helpers.ts`. Kept local so the shrunk inputs (trace params +
- * permutation seed) are sufficient to reproduce a failing case without
- * pulling in additional helpers.
- */
-const shuffleWithSeed = <T>(items: ReadonlyArray<T>, seed: number): T[] => {
-  let state = seed >>> 0;
-  const next = (): number => {
-    state = (state * 1_664_525 + 1_013_904_223) >>> 0;
-    return state / 0x1_0000_0000;
-  };
-  const result = items.slice();
-  for (let i = result.length - 1; i > 0; i--) {
-    const j = Math.floor(next() * (i + 1));
-    const tmp = result[i]!;
-    result[i] = result[j]!;
-    result[j] = tmp;
-  }
-  return result;
-};
 
 describe("property: delivery-order invariance for a fixed event set", () => {
   it("every permutation of the same event set yields the same final text", () => {
